Add PWM remap fonctionnality.
On GD32V103f there is 4 types of remaps for PWM. Not all timers are able to
be configured with this 4 remaps types. Also it is not possible to deduce the
Remap type with the pins used (e.g TIMER0).

To define a remap for PWM channels, you need to instantiate a PwmConfiguration
for a specific TIMER with a specific Remap type. From thin configuration, you
get channels corresponding to provided pins. And the call configure() to return
the PwmTimer type used to control the PWM.

```rust

    // ...

    let pwm_t0 = PwmConfiguration::<TIMER0, NoRemap>::new(timer0, &mut afio);
    let (_, ch1, ch2, _) = pwm_t0.pins(None, Some(&pa9), Some(&pa10), None);
    let mut pwm_t0 = pwm_t0.configure(&mut rcu);

    // ...
    pwm_t0.set_duty(ch1, 100);

```

It also simplify the PwmTimer `new` function. No more need to used the trait
`PwmChannelPin`.